Document the Jacobi preconditioner.
I am not sure about the API that should be presented to the programmer.
For instance, should `preconditioner` return the preconditioner `M`
or `M` inverse? With the Jacobi preconditioner the inverse is trivial,
so computing the inverse when solving is completely fine, but if we
implement some conditioner where matrix inversion is expensive...

There's also questions as how to structure left preconditioning, right
preconditioning, etc. and how that is best exposed to the programmer;
I feel fine about `solve` for `Mx = b` though (keep in mind the previous
paragraph).